walkback route change and init integ tests
