+++ app/integration/_index.test.ts
@@ -3,7 +3,7 @@ import { test, expect } from "@playwright/test";
 test("redirects to devices list", async ({ page }, testInfo) => {
   await page.goto("/", { waitUntil: "load" });
   // Wait for the redirect to complete
-  await page.waitForURL("**/devices", { timeout: 5000 });
-  
-  expect(page.url()).toBe(`http://localhost:50123/devices`);
+  await page.waitForURL("**/devices/list", { timeout: 5000 });
+
+  expect(page.url()).toBe(`http://localhost:50123/devices/list`);
 });
